refactor!: ImageIO dependency
imageio is moving towards v3 and throws deprecation warnings
since v2.16.2.
They introduced a v2 namespace to allow for using the old API in v2.16.0.
Since v2.16.0 does not affect any of our other dependencies
(we dropped Python 3.6 support either way),
I have updated the dependency and the import
as they suggested in the deprecation warning:

"Starting with ImageIO v3 the behavior of this function will
switch to that of iio.v3.imread.
To keep the current behavior (and make this warning dissapear)
use `import imageio.v2 as imageio` [...]."

See changelog of ImageIO for more info:
https://github.com/imageio/imageio/blob/master/CHANGELOG.md