fix: Create the correct anchors for Redmine TOC
Added a function that creates the HTML anchors for presentation objects
in the same manner as done by Redmine itself.
See #80 for a discussion of the issue.

Proposed solution: The function is passed to the Redmine template,
which can (but does not have to) use it to add a table of contents (TOC)

Additionally added a test that checks if the function works as expected.
I tried to be verbose in the documentation because Regex syntax is
powerful but not at all intuitive!

Additional changes come from black/isort