Redesign NEMO Timeseries, add proper unit conversion for global aggregation
- Add a new class NemoTimeseries that can be used for creating NEMO timeseries processing tasks
- make NemoGlobalMeanYearMeanTimeseries and NemoGlobalSumYearMeanTimeseries children of said class
- adjust entry points accordingly (user-facing behavior does not change)

Change spatial weights computation:
Iris can automatically take care of unit conversion when aggregating, as long as the spatial weights are added to the cube as a CellMeasure. (cf. https://scitools-iris.readthedocs.io/en/latest/userguide/cube_statistics.html#area-averaging).
I have thus changed the spatial_weights function to add a "cell_size" as a cell measure, containing the grid area/volume
This means I do not have to multiply by m2/m3 depending on the aggregation.
It also leads to a slight simplification of the SI3 Timeseries task.

Updated the tests, I also now check that the units are changed when summing (but not when computing the global mean)